backend/requirements.txt
streamlit>=1.32.0
python-dotenv>=1.0.1
requests==2.31.0
PyGithub>=1.59.0
pandas>=2.2.1
PyJWT==2.8.0
flask>=2.3.0
flask-cors>=4.0.0